+++ doc/whatsnew.rst
@@ -35,6 +35,7 @@ New features
 - Add :meth:`xarray.Dataset.hyoga.assign`.
 - Add :meth:`xarray.Dataset.hyoga.assign_icemask`.
 - Add :meth:`xarray.Dataset.hyoga.where_icemask`.
+- Method :meth:`xarray.Dataset.hyoga.interp` uses "land_ice_area_fraction".
 - Plot methods look for variable `land_ice_area_fraction`.
 
 Internal changes